# storage/openapi.yaml
openapi: 3.0.3
info:
  title: statekit storage API
  version: 0.3.0
  description: >
    Query API for stored statekit state documents, organized in three data
    layers. L1 (/state/summary, /state/targets) is the constantly polled
    current view without data payloads. L2 (/state/targets/{key},
    /state/states/{identity}) is the full current detail fetched on
    drill-down. L3 (/state/states/{identity}/timeline, /state/timeline,
    /state/timeline/bucket, incidents) is historical and explicitly bounded.
    Every layer GET sets a strong ETag and honors If-None-Match with 304, so
    an idle fleet polls as one small summary body plus 304s.
paths:
...
  /escalations/global:
    post:
      summary: Report a fleet-wide incident
      description: >
        Creates or updates a global incident such as a build, deployment, or
        rollback, without a component owning it. Omit id to create a new
        incident; include id (and optionally status closed) to annotate or
        close an existing one.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GlobalIncidentRequest'
      responses:
        '200':
          description: Updated incident
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
        '201':
          description: Created incident
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
  /escalations/ack:
    post:
      summary: Mark a stored escalation incident acknowledged
      parameters:
        - name: source
          in: query
          required: true
          schema:
            type: string
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Acknowledged
...
components:
...
  schemas:
...
    StateDisplayLabel:
      type: object
      required: [name, value]
      properties:
        name:
          type: string
        value:
          type: string
...
    Incident:
      type: object
      properties:
        identity:
          type: string
        source:
          type: string
        scraped_from:
          type: string
        scrape_path:
          type: string
        id:
          type: string
        type:
          type: string
          description: Incident type. Empty for a regular component incident; build, deployment, and rollback describe fleet-wide operations.
        title:
          type: string
        status:
          type: string
          enum: [open, closed, acknowledged]
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        last_updated_at:
          type: string
          format: date-time
        severity:
          type: string
          enum: [pass, warn, fail, down]
        label_path:
          type: array
          items:
            $ref: '#/components/schemas/StateDisplayLabel'
        labels:
          type: object
          additionalProperties:
            type: string
        topics:
          type: object
          additionalProperties: true
        events:
          type: array
          items:
            $ref: '#/components/schemas/IncidentEvent'
        observed_at:
          type: string
          format: date-time
    GlobalIncidentRequest:
      type: object
      properties:
        source:
          type: string
          description: Reporting source. Defaults to "global".
        id:
          type: string
          description: Existing incident ID to annotate or close. Omit to create.
        type:
          type: string
          description: Incident type, for example build, deployment, or rollback. Required on create.
        title:
          type: string
          description: Defaults to the type on create.
        severity:
          type: string
          enum: [pass, warn, fail, down]
        status:
          type: string
          description: Set to closed to close an existing incident.
          enum: [closed]
        message:
          type: string
          description: Event message recorded on the incident.
        labels:
          type: object
          additionalProperties:
            type: string
        data:
          type: object
          additionalProperties: true
    IncidentEvent:
      type: object
      properties:
        event_key:
          type: string
        identity:
          type: string
        seq:
          type: string
        timestamp:
          type: string
          format: date-time
        topic:
          type: string
        message:
          type: string
        data:
          type: object
          additionalProperties: true
        observed_at:
          type: string
          format: date-time